requirements.txt
replicate~=0.15.4
uvicorn==0.23.2
fastapi==0.103.0
pydantic==2.3.0
openai==0.28.0
langchain~=0.0.321
charset-normalizer==3.2.0
chardet==4.0.0
python-dotenv~=1.0.0